extracted home cards
