+++ client/src/controllers/Home/Home.js
@@ -1,53 +1,32 @@
 import React, {useEffect} from 'react'
 import './Home.scss'
 
-import Button from '../../components/Button/Button'
+import {connect} from 'react-redux'
+
 import Header from '../../components/Header/Header'
-import StoryCard from '../../components/StoryCard/StoryCard'
 import Modal from '../../components/Modal/Modal'
 import LogInRegisterModal from '../../components/LogInRegister/LogInRegister'
 
-import {connect} from 'react-redux'
-import {toggleModalAction} from '../../redux/actions/page'
 import {getAllStoriesAction} from '../../redux/actions/db_get'
 import ButtonsForHome from '../ButtonGroups/ButtonsForHome/ButtonsForHome'
+import HomeCards from './HomeCards'
 
 const Home = props => {
-    const {toggleModal, getAllStories} = props
-    const {user, modal, stories} = props
+    const {getAllStories} = props
+    const {modal, stories} = props
 
     useEffect(()=>{
         getAllStories()
     }, [getAllStories])
 
-    let cards = []
-
-    if(stories){
-        stories.forEach((story, i) => {
-            cards.push(
-                <StoryCard
-                    {...{
-                        key: i,
-                        story: story
-                    }}
-                />
-            )
-        })
-    }
-
-
     return (
     <div className='home'>
 
         {modal? <Modal> <LogInRegisterModal/> </Modal> : ''}
-
         <ButtonsForHome />
-
         <Header />
 
-        <div className='storyCards'>
-            {cards}
-        </div>
+        <HomeCards stories={stories} />
 
     </div>
     )
@@ -55,7 +34,6 @@ const Home = props => {
 
 const mapStateToProps = state => {
     return {
-        user: state.profile.user,
         modal: state.page.modal,
         stories: state.page.stories,
     }
@@ -63,7 +41,6 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
     return {
-        toggleModal: profile => dispatch(toggleModalAction()),
         getAllStories: () => dispatch(getAllStoriesAction()),
     }
 }
